requirements.txt
discord.py
transformers
torch
pandas
spacy
elasticsearch
chromadb
fastapi
pytest
pytorch-lightning